chore: migrate to new image name schema
The entrypoint has changed slightly. We now always assume that the
entrypoint is the slurm binary which the image is named. It is assumed
that the Slurm daemon will run in foreground mode.

Currently sackd lacks a foreground mode or option (e.g. '-D'), but it is
the responsibility of the image to promise OCI compliance. Hence no
longer needing to hack it in the helm chart.

(cherry picked from commit e0dd9937cb1db3d0c23860ac8af8f3c9e09fa8b4)